refactor: vendor the forked react-native-shadow-stack
