+++ src/components/token-family/TokenFamilyHeaderIcon.tsx
@@ -6,11 +6,11 @@ import { type Source } from 'react-native-fast-image';
 import { ImgixImage } from '@/components/images';
 import { Text } from '@/design-system';
 import * as i18n from '@/languages';
-import ShadowStack from '@/react-native-shadow-stack';
 import { borders } from '@/styles';
 import { type ThemeContextProps } from '@/theme/ThemeContext';
 import FallbackIcon from '@/utils/CoinIcons/FallbackIcon';
 import { initials } from '@/utils/formatters';
+import { ShadowStack } from '@/vendor/react-native-shadow-stack';
 
 import eyeSlash from '../../assets/sf-eye.slash.png';
