--- src/react-native-shadow-stack/index.js
@@ -1,3 +1,0 @@
-import ShadowStack from './ShadowStack';
-
-export default ShadowStack;
