+++ tsconfig.json
@@ -22,7 +22,6 @@
       "logger": ["src/utils/logger"],
       "react-native-cool-modals": ["src/react-native-cool-modals"],
       "@rainbow-me/animated-charts": ["src/react-native-animated-charts/src"],
-      "react-native-shadow-stack": ["src/react-native-shadow-stack"],
       "@nktkas/hyperliquid/api/info": ["node_modules/@nktkas/hyperliquid/esm/api/info/mod"],
       "@nktkas/hyperliquid/api/exchange": ["node_modules/@nktkas/hyperliquid/esm/api/exchange/mod"],
       "@nktkas/hyperliquid/api/subscription": ["node_modules/@nktkas/hyperliquid/esm/api/subscription/mod"],
